impak-decoder/pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "impakdecoder"
version = "0.1.1"
description = "A decoder for image collection format impak"
readme = "README.md"
license = "AGPL-3.0-or-later"
license-files = ["LICENSE"]
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0",
]

[tool.setuptools.packages.find]
where = ["src"]